chore: rebuild frontend assets
