+++ frontend/src/layouts/JobsLayout.module.css
@@ -1,17 +1,9 @@
 .header {
   display: flex;
-  align-items: center;
-  justify-content: space-between;
+  justify-content: flex-end;
   padding: var(--spacing-lg);
   background: var(--color-surface);
   border-bottom: 1px solid var(--color-border);
-}
-.search {
-  padding: var(--spacing-xs) var(--spacing-md);
-  border: 1px solid var(--color-border);
-  border-radius: var(--radius-md);
-  width: 100%;
-  max-width: 20rem;
 }
 .addButton {
   padding: var(--spacing-sm) var(--spacing-md);
