+++ frontend/src/pages/JobsDiscoverPage.module.css
@@ -31,6 +31,13 @@
   gap: var(--spacing-md);
   margin-bottom: var(--spacing-xl);
 }
+.search {
+  padding: var(--spacing-xs) var(--spacing-md);
+  border: 1px solid var(--color-border);
+  border-radius: var(--radius-md);
+  width: 100%;
+  max-width: 20rem;
+}
 .filterRow {
   display: flex;
   flex-wrap: wrap;
